+++ src/components/BarraBusqueda.css
@@ -11,7 +11,9 @@
     width: 320px;
     height: 36px;
     background-color: #E7E7E7;
-    margin: 10px;
+    margin-left: 24px;
+    margin-top: 10px;
+    margin-bottom: 4px;
     }
 
 .barra-busqueda-input {
@@ -24,7 +26,6 @@
     letter-spacing: 1.59px;
     color: black;
     font-weight: bold;
-    margin: auto;
 }
 
 .barra-busqueda-input::placeholder {
