+++ src/components/BarraPresupuesto.css
@@ -1,7 +1,10 @@
 .barra-presupuesto-container{
     display: flex;
     justify-content: center; 
-    margin: 10px;
+    margin-left: 24px;
+    margin-top: 4px;
+    margin-bottom: 4px;
+    align-items: center;
 }
 
 .barra-presupuesto-input {
