+++ src/components/BarrasContainer.css
@@ -1,0 +1,5 @@
+.barras-container {
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+}
